requirements.txt
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.40
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1